Divide tests between feel and formula specific tests
